requirements.txt
scipy
numpy
opencv-python
matplotlib
pycocotools
pytorch>=1.0.1
filterpy==1.4.1
numba==0.38.1
scikit-image==0.14.0
scikit-learn==0.19.1
